# Scroll Viewer rendering optimization