--- a/gui/src/2D/controls/container.ts
+++ b/gui/src/2D/controls/container.ts
@@ -304,7 +304,7 @@ export class Container extends Control {
             return false;
         }
 
-        Control.numLayoutCalls++;
+        Control.NumLayoutCalls++;
 
         if (this._isDirty) {
             this._currentMeasure.transformToRef(this._transformMatrix, this._prevCurrentMeasureTransformedIntoGlobalSpace);