--- a/gui/src/2D/controls/scrollViewers/scrollViewerWindow.ts
+++ b/gui/src/2D/controls/scrollViewers/scrollViewerWindow.ts
@@ -37,7 +37,7 @@ export class _ScrollViewerWindow extends Container {
 
         var measure = new Measure(0, 0, renderWidth, renderHeight);
 
-        Control.numLayoutCalls = 0;
+        Control.NumLayoutCalls = 0;
 
         this.host._rootContainer._layout(measure, context);