Addressing #7341 issue.

I tried to not change the codepath when the optimization is disabled, to be sure that if something is broken it's only when enabling the optimization.

So, I added the `ScrollViewer.freezeControls` property.

When set to `true`, the `layout` method is not called anymore for the window child controls, and only the controls intersecting the window are rendered. Instead of calling `layout` for the children, I simply update the top/left properties according to the window translation values.

I have also added the `ScrollViewer.forceHorizontalBar` and `ScrollViewer.forceVerticalBar` properties. When set to `true`, they force the display of the corresponding bars. When you know your scroll viewer will end up with visible bars, you can set those properties to `true` to save some initialization time, as if it is the scroll viewer control that makes a bar visible in the course of the initialization, it will trigger a children layout rebuild, adding more time to the initialization process.

I have also modified `Measure.transformToRef` to get rid of the `new` calls because it led to some major GCs in my testings.

Knowing the number of `layout` / `render` calls came in handy when developing this, so I kept the feature: let me know if I should remove it.

Samples:
* https://popov72.github.io/BabylonDev/showSample.html?sample=scrollviewer2 - 20000 controls. On my computer, it runs around 2 fps when controls are not frozen and 60 fps when they are.
* https://popov72.github.io/BabylonDev/showSample.html?sample=scrollviewer - 276052 controls (it takes some time before displaying something because of the 276000 control creation process). It's the original sample from @salamaashoush in #7341. It runs around 0.5 fps when controls are not frozen and 12 fps when they are. Note that for this one we can see that there are some renderings with only 200/300 rendering calls for a frame whereas it should always be 276000+. I don't understand why, in the previous sample we always have 20000+ renderings for each frame as expected...

Note that I had to change `protected` to `public` for the `Control._children` property because I needed to traverse the control hierarchy in the exact same way than the `layout` and `render` functions, and using the getter `Control.children` was not possible because the grid control (maybe others too) overrides it and returns only the children of the cells and not the cells themselves.