@Popov72 2020-01-27
Added a new optimization (`setBucketSizes(width, height)`), [sample 2](https://popov72.github.io/BabylonDev/showSample.html?sample=scrollviewer) is now at 60 fps too.

From the doc:

Sets the bucket sizes.
When `freezeControls` is true, setting a non-zero bucket size will improve performances by updating only controls that are visible. The bucket sizes is used to subdivide (internally) the window area to smaller areas into which controls are dispatched. So, the size should be roughly equals to the mean size of all the controls of the window. To disable the usage of buckets, sets either width or height (or both) to 0.
Please note that using this option will raise the memory usage (the higher the bucket sizes, the less memory used), that's why it is not enabled by default.